Verboser error message
Spell out the problem in the error message for user-friendliness.

Motivation: When this error is raised while Rails is rendering a view,
Rails helpfully rescues it and re-raises it as an
`ActionView::Template::Error`. So, I was getting just
`ActionView::Template::Error: 0`, which is less than helpful.

Also, rename `LengthNotOne` to `CountNotOne` since we use the `count`
method and not the `length` method.